% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miguelcltb.R
\name{miguelcltb}
\alias{miguelcltb}
\title{miguelcltb}
\usage{
miguelcltb(n, iter, p = 0.5, ...)
}
\arguments{
\item{n}{sample size}

\item{iter}{number of iterations}

\item{p}{probability}

\item{...}{other arguments}
}
\value{
histogram
}
\description{
miguelcltb
}
\examples{
miguelcltb(n=5,iter=10000,p=0.5)

}
