% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Myboot2}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of iterations}

\item{x}{population}

\item{fun}{desired statistic}

\item{alpha}{confidence interval}

\item{cx}{scale}

\item{...}{alternate arguments}
}
\value{
histogram with confidence interval
}
\description{
Myboot2
}
\examples{
myboot2(x=c(1,3,4,5,3,2), fun="mean", alpha=0.05)

}
