% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myddt1.R
\name{myddt1}
\alias{myddt1}
\title{myddt1}
\usage{
myddt1(df, x, y, SPECIES, col)
}
\arguments{
\item{df}{data}

\item{x}{variable}

\item{y}{variable}

\item{SPECIES}{conditional variable}

\item{col}{color of river}
}
\value{
Plot of LENGTH vs WEIGHT
}
\description{
myddt1
}
\examples{
\dontrun{myddt1(df=ddt, SPECIES="CCATFISH", x = "LENGTH", y= "WEIGHT", col = "RIVER")}
}
