% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve}
\usage{
myncurve(mu, sigma, x)
}
\arguments{
\item{mu}{mean}

\item{sigma}{standard deviation}

\item{x}{limit of x=a of mu +/- 3sigma}
}
\value{
graph of distribution with labeling
}
\description{
myncurve
}
\examples{
\dontrun{(mu=5,sigma=3,6)}
}
