% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mynormfunction.R
\name{mynormfunction}
\alias{mynormfunction}
\title{My norm function}
\usage{
mynormfunction(x, mu, sig, ...)
}
\arguments{
\item{x}{data to be passed through}

\item{mu}{mean}

\item{sig}{standard deviation}

\item{...}{alternate arguments}
}
\value{
graph of maximum likelihood
}
\description{
My norm function
}
\examples{
mynormfunction(x=c(5,7,7,8,10),mu=seq(5,10,length=1000),sig=seq(0.1,4,length=1000),lwd=2,labcex=1)
}
