% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterhist.R
\name{scatterhist}
\alias{scatterhist}
\title{Title Scatter Histogram}
\usage{
scatterhist(x, y, xlab = "", ylab = "")
}
\arguments{
\item{x}{vector}

\item{y}{vector}

\item{xlab}{string}

\item{ylab}{string}
}
\value{

}
\description{
Title Scatter Histogram
}
\examples{
scatterhist(c(1,5,8,7),c(7,2,9,3),"x-axis","y-axis")
}
